requirements.txt
torch
torchvision
transformers
medmnist
numpy
pillow
matplotlib
scikit-learn
lime
shap
scikit-image
tqdm
